Move legend click handler to separate file
